Refactored assembler.rs to parser.rs
